requirements.txt
flask
pymongo
flask_cors
jwt
werkzeug.security
datetime
dotenv
os
re
functools